Make utils corotine scope
