+++ ktor-client/ktor-client-core/ktor-client-core-js/src/io/ktor/client/engine/js/compatible/Utils.kt
@@ -2,13 +2,16 @@ package io.ktor.client.engine.js.compatible
 
 import io.ktor.client.engine.js.compatible.browser.*
 import io.ktor.client.engine.js.compatible.node.*
+import kotlinx.coroutines.*
 import kotlinx.coroutines.io.*
 import org.w3c.fetch.*
 import kotlin.coroutines.*
 import kotlin.js.*
 
 
-abstract class Utils {
+abstract class Utils : CoroutineScope {
+    override val coroutineContext: CoroutineContext = SupervisorJob()
+
     companion object {
         fun get(): Utils {
             return if (hasFetchApi()) {
